printf size_t fix for 32bit.
There's no real way to print a size_t, but we know it probably
everywhere fits in a 64bit unsigned integer on all arches we
currently support, so cast it to an uint64_t and print as %llu
should work everywhere.

%zu is C99, we could potentially use that too, yikes!

+++ test/check_daemon.c
@@ -50,7 +50,7 @@ static char* dump_fuzz(FuzzContext *fuzz)
 
 	fprintf(s, "\n\n******************************************\n");
 	fprintf(s, "current time %ld\n", time(NULL));
-	fprintf(s, "iteration: %d\tsize: %ld\n", fuzz->iteration, fuzz->size);
+	fprintf(s, "iteration: %d\tsize: %llu\n", fuzz->iteration, (unsigned long long)fuzz->size);
 	for (int i = 0; i < fuzz->size; i++) {
 		fprintf(s, "%02X ", fuzz->buf[i]);
 		c+= 3;
